fix enable strict mode for CI
By default, the pipefail is disabled. And Git
For instance, the commands should
return 1 as exit code, but we get zero. After pipefail enabled, it's
back to normal.

```bash
$ ( echo hello; exit 1 ) | tee 2>&1
hello
$ echo $?
0

$ set -o pipefail
$ ( echo hello; exit 122 ) | tee 2>&1
hello
$ echo $?
122
```

And the test.log is used to grep the error from a ton of messages. So
that we should `set +e` and set it back after egrep.

Signed-off-by: Wei Fu <fuweid89@gmail.com>